g4-common: update radio blobs to v29a
In this commit, the included files were taken from LG G4 H815 v29a, and
g4-common-vendor.mk updated to account for the new files.

Most files are unmodified from v29a, except the below files were patched:

	proprietary/lib/libprotobuf-cpp-full.so
	proprietary/lib/libprotobuf-cpp-lite.so
	proprietary/lib64/libprotobuf-cpp-full.so
	proprietary/lib64/libprotobuf-cpp-lite.so
	proprietary/vendor/lib/libcne.so
	proprietary/vendor/lib/libcneapiclient.so
	proprietary/vendor/lib/libril-qc-qmi-1.so
	proprietary/vendor/lib/libsettings.so
	proprietary/vendor/lib/libwms.so
	proprietary/vendor/lib/libwqe.so
	proprietary/vendor/lib64/libcne.so
	proprietary/vendor/lib64/libcneapiclient.so
	proprietary/vendor/lib64/libril-qc-qmi-1.so
	proprietary/vendor/lib64/libsettings.so
	proprietary/vendor/lib64/libwms.so
	proprietary/vendor/lib64/libwqe.so

Patches to the "libril-qc-qmi-1.so" files were manually done in a hex editor:
  "/system/lib/libvss_common_core.so" --> "/vendor/lib/libvss_common_core.so"
  "/system/lib/libvss_nv_core.so" --> "/vendor/lib/libvss_nv_core.so"
  "/system/bin/diag_mdlog" --> "/vendor/bin/diag_mdlog"
...and similiarly for /system/lib64 to /vendor/lib64.

The remaining patches were made with "patchelf 0.9". These patches allow us
to use both the AOSP and vendor "libprotobuf" side-by-side. The vendor
"libprotobuf" was renamed to "libprotobuf-29a" and all references updated to
match it.

The patches were performed by the following script:

#!/bin/bash
# Patch libraries to rename from "libprotobuf-cpp-*" to "libprotobuf-cpp-*-29a".
for FILE in lite full; do
    patchelf --set-soname libprotobuf-cpp-$FILE-29a.so proprietary/lib64/libprotobuf-cpp-$FILE.so
    patchelf --set-soname libprotobuf-cpp-$FILE-29a.so proprietary/lib/libprotobuf-cpp-$FILE.so
done

# Patch libraries to reference "libprotobuf-cpp-lite-29a" instead of "libprotobuf-cpp-lite"
for FILE in cne cneapiclient wms wqe; do
	patchelf --replace-needed libprotobuf-cpp-lite.so libprotobuf-cpp-lite-29a.so proprietary/vendor/lib64/lib$FILE.so
	patchelf --replace-needed libprotobuf-cpp-lite.so libprotobuf-cpp-lite-29a.so proprietary/vendor/lib/lib$FILE.so
done

# Patch libraries to reference "libprotobuf-cpp-full-29a" instead of "libprotobuf-cpp-full"
for FILE in settings; do
	patchelf --replace-needed libprotobuf-cpp-full.so libprotobuf-cpp-full-29a.so proprietary/vendor/lib64/lib$FILE.so
	patchelf --replace-needed libprotobuf-cpp-full.so libprotobuf-cpp-full-29a.so proprietary/vendor/lib/lib$FILE.so
done
# End script

N.B. In LineageOS 18.0, "patchelf" is a prebuilt binary, thus these patches
may be done at compile-time. If you're from the future and are reading this,
it is your job to port these changes to LineageOS 18.0. You can do it!